refactor(search): merge single and multi field topn plan visitors
Both visitors emitted the same IndexOptimizeMode::SimpleTopN and were
structurally identical; fold them into one SimpleTopnVisitor handling
1..=MAX_SIMPLE_TOPN_FIELDS group by fields.

The merged visitor applies the secondary-sort validation (must be ASC
columns referencing the group by fields) to all arities — previously the
multi-field path accepted arbitrary secondary sorts unvalidated. It also
enforces projection len == group fields + 1 for every arity, replacing
the single-field-only exact-2 check.
