+++ src/service/search/datafusion/optimizer/physical_optimizer/index_optimizer/mod.rs
@@ -39,7 +39,6 @@ mod distinct;
 mod histogram;
 mod select;
 mod topn;
-mod topn_multi;
 mod utils;
 
 use crate::service::search::datafusion::{
@@ -50,7 +49,6 @@ use crate::service::search::datafusion::{
         histogram::{is_simple_histogram, is_simple_multi_histogram},
         select::is_simple_select,
         topn::is_simple_topn,
-        topn_multi::is_simple_topn_multi,
         utils::is_complex_plan,
     },
 };
@@ -147,15 +145,10 @@ impl TreeNodeRewriter for FollowerIndexOptimizer {
                 return Ok(Transformed::new(plan, true, TreeNodeRecursion::Stop));
             }
 
-            // check if the query is simple topn, topn_multi or simple distinct
+            // check if the query is simple topn or simple distinct
             if config::cluster::LOCAL_NODE.is_single_node() {
                 if let Some(index_optimize_mode) =
                     is_simple_topn(Arc::clone(&plan), self.index_fields.clone())
-                {
-                    *self.index_optimizer_mode.lock() = Some(index_optimize_mode);
-                    return Ok(Transformed::new(plan, true, TreeNodeRecursion::Stop));
-                } else if let Some(index_optimize_mode) =
-                    is_simple_topn_multi(Arc::clone(&plan), self.index_fields.clone())
                 {
                     *self.index_optimizer_mode.lock() = Some(index_optimize_mode);
                     return Ok(Transformed::new(plan, true, TreeNodeRecursion::Stop));
@@ -273,13 +266,6 @@ impl TreeNodeRewriter for LeaderIndexOptimizer {
                 is_simple_topn(Arc::clone(&plan), index_fields.clone())
             {
                 // Check for SimpleTopN
-                let mut rewriter = IndexOptimizerRewrite::new(index_optimize_mode);
-                let plan = plan.rewrite(&mut rewriter)?.data;
-                return Ok(Transformed::new(plan, true, TreeNodeRecursion::Stop));
-            } else if let Some(index_optimize_mode) =
-                is_simple_topn_multi(Arc::clone(&plan), index_fields.clone())
-            {
-                // Check for multi-field GROUP BY top-n
                 let mut rewriter = IndexOptimizerRewrite::new(index_optimize_mode);
                 let plan = plan.rewrite(&mut rewriter)?.data;
                 return Ok(Transformed::new(plan, true, TreeNodeRecursion::Stop));
